refactor(maestro): split ide/completion root into cohesive modules
